(vm)splice: add SPLICE_F_NOSIGPIPE
This flag has the same behavior as MSG_NOSIGNAL has for sockets:
prevents the emission of SIGPIPE to the process trying to write to a
pipe with no reader. The caller still gets -EPIPE.

In order to achieve that, I had to makes splice_to_pipe take a flags
argument, the same one that the .splice_read functions get.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>