Introduce O_NOSIGPIPE for pipes and FIFOs
This flag can be passed to open(2) and pipe2(2) to make the writing end
not send SIGPIPE when a sender tries to write to a pipe with no
readers. It can also be passed to fcntl(2) to set/unset the flag after
the pipe is opened. With this flag, the sender will still receive -EPIPE
error indicating the pipe is closed. This is like MSG_NOSIGNAL for
sendmsg(2) but is sticky: it applies to all writes done on that pipe.

This flag is important for applications dealing with multiple pipes, as
more often than not they don't want to be terminated when one of the
pipes closes. The only two pipes one usually expects to terminate the
application are stdout and stderr. However, without this flag, those
applications need to always set SIGPIPE to SIG_IGN, which makes SIGPIPE
not be delivered for stdout and stderr either.

On the name: I chose xxx_NOSIGPIPE instead of xxx_NOSIGNAL (à la
MSG_NOSIGNAL) because those names already exist on NetBSD[1][2] and some
do on OS X[3] and FreeBSD[4].

[1] http://netbsd.gw.com/cgi-bin/man-cgi?open
[2] https://mail-index.netbsd.org/tech-kern/2012/01/24/msg012535.html
[3] https://developer.apple.com/library/mac/documentation/Darwin/Reference/ManPages/man2/fcntl.2.html
[4] http://www.freebsd.org/cgi/man.cgi?sektion=2&query=setsockopt

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>