Add SOCK_NOSIGPIPE to prevent sockets from sending SIGPIPE
Like O_NOSIGPIPE for pipes and MSG_NOSIGNAL in each sendto/sendmsg call,
this makes it so a send on a socket that disconnected will not produce
SIGPIPE back to the caller. The implementation simply sets the
MSG_NOSIGNAL upon entry to sendto(2) and sendmsg(2), just like
MSG_DONTWAIT is set if the socket has O_NONBLOCK set.

The name matches the flag in NetBSD[1].

[1] http://netbsd.gw.com/cgi-bin/man-cgi?socket

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>